requirements.txt
fastapi
fastapi-cli 
uvicorn[standard]
tortoise-orm
PyJWT
slowapi
pydantic
pydantic-settings
aiosqlite
pytest
pytest-cov
pytest-asyncio 
httpx==0.27.2
ruff 
black 
mypy 